backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
googlemaps>=4.10.0
pydantic-settings>=2.0.3
